Bundle Changes
Handled all sort of stuff to get the WaveManager to start working. Such as:
Created variable in WaveManager which each Level can access to give it a maximum of waves.

Handled wave checking so that new waves spawn when the old one is completed, also handles results when max waves is hit.